Reorganize annotation tool file layout
