+++ web_external/routes.js
@@ -225,7 +225,7 @@ import TasksView from './Tasks/TasksView';
 router.route('tasks', 'tasks', () => {
     navigateToIfLoggedIn(TasksView);
 });
-import AnnotationTool from './vue/views/AnnotationTool.vue';
+import AnnotationTool from './vue/components/AnnotationTool/AnnotationTool.vue';
 router.route('tasks/annotate/:id', 'annotate', (id) => {
     navigateToIfLoggedIn(VueComponentView, {
         component: AnnotationTool,
