+++ web_external/vue/components/AnnotationTool/AnnotationToolFeatures.vue
