+++ web_external/vue/components/AnnotationTool/AnnotationToolStore.js
